app/src/main/java/com/example/anuraj/tutorialsapplication/NetworkUtils.java
package com.example.anuraj.tutorialsapplication;

import android.content.Context;
import android.net.ConnectivityManager;
import android.net.NetworkInfo;

/**
 * Class with method(s) to check network connectivity
 */

public class NetworkUtils {

    /**
     * Checks internet connectivity
     *
     * @param context Activity context
     * @return true / false
     */
    public static boolean isNetworkAvailable(Context context) {
        ConnectivityManager connectivityManager
                = (ConnectivityManager) context.getSystemService(Context.CONNECTIVITY_SERVICE);
        NetworkInfo activeNetworkInfo = connectivityManager.getActiveNetworkInfo();
        return activeNetworkInfo != null && activeNetworkInfo.isConnectedOrConnecting();

    }



}
